TD/reponses/sources/prediction/requirements.txt
google-cloud-bigquery
pandas
db-dtypes
